refactor buidpack detection to work for any directory
